# static/swagger.yml
openapi: 3.0.0
paths:
...
  /api/v2/crontab/create/users:
    post:
      operationId: crontab-create-users
      summary: Включение CronTab для пользователей (только администратор)
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CrontabCreateRequest"
      responses:
        "200":
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400": &ref_109
          description: Ответ будет таким если с данным что-то не так
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadRequestError"
        "401": &ref_110
          description: Ответ для незарегистрированного пользователя
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
        "403": &ref_111
          description: Ответ для неавторизованного пользователя
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForbiddenError"
        "404": &ref_112
          description: Не найдено
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
        "406": &ref_113
          description: Не принято значение
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotAcceptableError"
        "409": &ref_114
          description: Ответ для конфликта файлов
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConflictError"
        "412": &ref_115
          description: Пользователь уже существует
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PreconditionFailedError"
        "500": &ref_116
          description: Ошибка сервера
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerError"
        "501": &ref_117
          description: Пока не реализовано
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotImplementedError"
      tags: &ref_118
        - crontab
      security: &ref_119
        - bearer: []
  /api/v2/crontab/delete/users:
    post:
      operationId: crontab-delete-users
      summary: Выключение CronTab для пользователей (только администратор)
      parameters: []
      responses:
        "200":
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400": *ref_109
        "401": *ref_110
        "403": *ref_111
        "404": *ref_112
        "406": *ref_113
        "409": *ref_114
        "412": *ref_115
        "500": *ref_116
        "501": *ref_117
      tags: *ref_118
      security: *ref_119
info:
  title: MyScreen backend API
  description: MyScreen backend API
  version: 2.41.29
  contact:
    name: Stanislav V Vyaliy
    url: https://wisekaa.dev
    email: stanislav@wisekaa.dev
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
      description: Токен авторизации
      name: token
  schemas:
    BadRequestError:
      type: object
      properties:
        statusCode:
          type: number
          example: 400
        code:
          type: string
          example: server-error.10004
        message:
          type: string
          example: Bad request
      required:
        - statusCode
        - code
        - message
    UnauthorizedError:
      type: object
      properties:
        statusCode:
          type: number
          example: 401
        code:
          type: string
          example: server-error.10001
        message:
          type: string
          example: Unauthorized request
      required:
        - statusCode
        - code
        - message
    ForbiddenError:
      type: object
      properties:
        statusCode:
          type: number
          example: 403
        code:
          type: string
          example: server-error.10002
        message:
          type: string
          example: Forbidden
      required:
        - statusCode
        - code
        - message
    NotFoundError:
      type: object
      properties:
        statusCode:
          type: number
          example: 404
        code:
          type: string
          example: server-error.10005
        message:
          type: string
          example: Not Found
      required:
        - statusCode
        - code
        - message
    NotAcceptableError:
      type: object
      properties:
        statusCode:
          type: number
          example: 406
        code:
          type: string
          example: server-error.10000
        message:
          type: string
          example: Not Acceptable
      required:
        - statusCode
        - code
        - message
    FileIDResponse:
      type: object
      properties:
        id:
          type: string
          description: Идентификатор файла
          format: uuid
        name:
          type: string
          description: Имя файла
          example: foo.mp4
      required:
        - id
        - name
    ConflictDataFile:
      type: object
      properties:
        id:
          type: string
          description: ID ссылки
          format: uuid
        name:
          type: string
          description: Имя ссылки
          example: Имя ссылки
        file:
          description: Файл
          allOf:
            - $ref: "#/components/schemas/FileIDResponse"
    ConflictData:
      type: object
      properties:
        video:
          description: Редакторы (видео)
          allOf:
            - $ref: "#/components/schemas/ConflictDataFile"
        audio:
          description: Редакторы (аудио)
          allOf:
            - $ref: "#/components/schemas/ConflictDataFile"
        playlist:
          description: Плэйлисты
          allOf:
            - $ref: "#/components/schemas/ConflictDataFile"
        monitor:
          description: Мониторы
          allOf:
            - $ref: "#/components/schemas/ConflictDataFile"
    ConflictError:
      type: object
      properties:
        statusCode:
          type: number
          example: 409
        code:
          type: string
          example: server-error.10000
        message:
          type: string
          example: Conflict exists
        data:
          $ref: "#/components/schemas/ConflictData"
      required:
        - statusCode
        - code
        - message
        - data
    PreconditionFailedError:
      type: object
      properties:
        statusCode:
          type: number
          example: 412
        code:
          type: string
          example: server-error.10002
        message:
          type: string
          example: User exists
      required:
        - statusCode
        - code
        - message
    InternalServerError:
      type: object
      properties:
        statusCode:
          type: number
          example: 500
        code:
          type: string
          example: server-error.10000
        message:
          type: string
          example: Server error
      required:
        - statusCode
        - code
        - message
    NotImplementedError:
      type: object
      properties:
        statusCode:
          type: number
          example: 501
        code:
          type: string
          example: server-error.10000
        message:
          type: string
          example: Server error
      required:
        - statusCode
        - code
        - message
    Status:
      type: string
      enum:
        - success
        - error
...
    SuccessResponse:
      type: object
      properties:
        status:
          description: Статус операции
          example: success
          $ref: "#/components/schemas/Status"
      required:
        - status
...
    CrontabCreateRequest:
      type: object
      properties:
        crontab:
          type: string
          description: "Crontab: \"Second Minute Hour DayOfMonth Month DayOfWeek\""
          example: 0 0 0 * * *
...